move pages directory to the correct location
