+++ frontend/src/router/Router.tsx
@@ -1,6 +1,6 @@
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import App from '../App';
-import NotFound from '../components/pages/NotFound';
+import NotFound from '../pages/NotFound';
 
 const routes = createBrowserRouter([
   {
